(Traffic BLX) Updated the ComplexPhases domain.rddl
