+++ rddlrepository/archive/standalone/TrafficBLX/ComplexPhases/__init__.py
@@ -1,0 +1,7 @@
+info = {
+    'name': 'TrafficBLX_ComplexPhases',
+    'description': 'BLX/QTM traffic signal control model with a generic phasing scheme. The goal is to control traffic lights to minimize total travel time.',
+    'context': '',
+    'tags': '',
+    'viz': 'TrafficViz.TrafficVisualizer'
+}
